Fixed mac compilation warnings.

+++ winpr/libwinpr/library/library.c
@@ -251,7 +251,7 @@ DWORD GetModuleFileNameA(HMODULE hModule, LPSTR lpFilename, DWORD nSize)
 
 #elif defined(__MACOSX__)
 	int status;
-	int length;
+	size_t length;
 
 	if (!hModule)
 	{
@@ -278,7 +278,7 @@ DWORD GetModuleFileNameA(HMODULE hModule, LPSTR lpFilename, DWORD nSize)
 		{
 			CopyMemory(lpFilename, buffer, length);
 			lpFilename[length] = '\0';
-			return length;
+			return (DWORD)length;
 		}
 
 		CopyMemory(lpFilename, buffer, nSize - 1);
